requirements.txt
torch>=2.2
torchaudio>=2.2
numpy>=1.24
pandas>=2.0
pyyaml>=6.0
tqdm>=4.66
scikit-learn>=1.4
matplotlib>=3.8
